% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cokm.param.R
\name{cokm.param}
\alias{cokm.param}
\title{Get model parameters in the autoregressive cokriging model}
\usage{
cokm.param(obj)
}
\arguments{
\item{obj}{a \code{\link{cokm}} object construted via the function \code{\link{cokm}} in 
this package}
}
\value{
a list of model parameters including regression coefficients \eqn{\beta}, 
scale discrepancy \eqn{\gamma}, variance parameters 
\eqn{\sigma^2}, and correlation parameters \eqn{\phi} in covariance functions.
If nugget parameters are included in the model, then nugget parameters are shown in \eqn{\phi}.
}
\description{
This function compute estimates for regression and variance
 parameters given the correlation parameters are known. It is used to show
all model parameters in one place.
}
\seealso{
\code{\link{cokm}}, \code{\link{cokm.fit}}, \code{\link{cokm.condsim}}, \code{\link{ARCokrig}}
}
\author{
Pulong Ma <mpulong@gmail.com>
}
