% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcokm.fit.R
\name{mvcokm.fit}
\alias{mvcokm.fit}
\title{fit the autoregressive cokriging model for multivariate output}
\usage{
mvcokm.fit(obj)
}
\arguments{
\item{obj}{a \code{\link{mvcokm}} object construted via the function \code{\link{mvcokm}} in 
this package}
}
\description{
This function estimates parameters in
 the parallel partial cokriging model
}
\seealso{
\code{\link{mvcokm}}, \code{\link{mvcokm.predict}}, \code{\link{mvcokm.condsim}}, \code{\link{ARCokrig}}
}
\author{
Pulong Ma <mpulong@gmail.com>
}
