% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcokm.param.R
\name{mvcokm.param}
\alias{mvcokm.param}
\title{Get model parameters in autoregressive cokriging models for multivarite output}
\usage{
mvcokm.param(obj)
}
\arguments{
\item{obj}{a \code{\link{mvcokm}} object construted via the function \code{\link{mvcokm}} in 
this package}
}
\value{
a list of model parameters including regression coefficients \eqn{\beta}, 
scale discrepancy \eqn{\gamma}, variance parameters 
\eqn{\sigma^2}, and correlation parameters \eqn{\phi} in covariance functions.
If nugget parameters are included in the model, then nugget parameters are shown in \eqn{\phi}.
}
\description{
This function computes estimates for regression and variance parameters
 given the correlation parameters are known. It is used to show all model
parameters in one place.
}
\seealso{
\code{\link{mvcokm}}, \code{\link{mvcokm.fit}}, \code{\link{mvcokm.predict}}, \code{\link{ARCokrig}}
}
\author{
Pulong Ma <mpulong@gmail.com>
}
