% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcokm.predict.R
\name{mvcokm.predict}
\alias{mvcokm.predict}
\title{Prediction at new inputs in autoregressive cokriging models for multivarite output}
\usage{
mvcokm.predict(obj, input.new)
}
\arguments{
\item{obj}{a \code{\link{mvcokm}} object construted via the function \code{\link{mvcokm}} in 
this package}

\item{input.new}{a matrix including new inputs for making prediction}
}
\description{
This function makes prediction in
 the parallel partial cokriging model. If a nested design is used, the predictive mean and predictive variance are
 computed exactly; otherwise, Monte Carlo simulation from the predictive distribution is used to approximate
the predictive mean and predictive variance.
}
\seealso{
\code{\link{mvcokm}}, \code{\link{mvcokm.fit}}, \code{\link{mvcokm.condsim}}, \code{\link{ARCokrig}}
}
\author{
Pulong Ma <mpulong@gmail.com>
}
